issue #16 fix types, add styles
