+++ src/js/component/PositionAdd/PositionAdd.scss
@@ -1,0 +1,5 @@
+.position-add {
+  display: flex;
+  margin: 10px 0;
+  max-width: 320px;
+}
